[ISSUE #6301] move the package and rename the Selector same to 1.x.
