+++ naming/src/main/java/com/alibaba/nacos/naming/misc/UtilsAndCommons.java
@@ -20,8 +20,8 @@ import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.api.selector.SelectorType;
 import com.alibaba.nacos.common.utils.JacksonUtils;
 import com.alibaba.nacos.common.utils.VersionUtils;
-import com.alibaba.nacos.naming.selector.LabelSelector;
-import com.alibaba.nacos.naming.selector.NoneSelector;
+import com.alibaba.nacos.naming.selector.v1.LabelSelector;
+import com.alibaba.nacos.naming.selector.v1.NoneSelector;
 import com.alibaba.nacos.sys.env.EnvUtil;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.alibaba.nacos.common.utils.StringUtils;
