--- naming/src/main/java/com/alibaba/nacos/naming/selector/CmdbLabelSelector.java
@@ -1,131 +1,0 @@
-/*
- *  Copyright 1999-2021 Alibaba Group Holding Ltd.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.alibaba.nacos.naming.selector;
-
-import com.alibaba.nacos.api.cmdb.pojo.Entity;
-import com.alibaba.nacos.api.exception.NacosException;
-import com.alibaba.nacos.api.naming.pojo.Instance;
-import com.alibaba.nacos.api.naming.selector.AbstractCmdbSelector;
-import com.alibaba.nacos.api.naming.selector.Selector;
-import com.alibaba.nacos.api.naming.selector.context.CmdbContext;
-import com.alibaba.nacos.common.utils.CollectionUtils;
-import com.alibaba.nacos.common.utils.StringUtils;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-/**
- * The implement of {@link LabelSelector} at new version. The main logic is same with {@link LabelSelector}.
- * The {@link CmdbLabelSelector} will return the instances labels in {@link #labels} and providers' label value is same with consumer.
- * If none matched, then will return all providers instead of.
- *
- * @author chenglu
- * @date 2021-07-16 16:26
- */
-public class CmdbLabelSelector<T extends Instance> extends AbstractCmdbSelector<T> {
-    
-    private static final String TYPE = "label";
-    
-    /**
-     * {@link Entity} labels key.
-     */
-    private Set<String> labels;
-    
-    /**
-     * the labels expression.
-     */
-    private String expression;
-    
-    public Set<String> getLabels() {
-        return labels;
-    }
-    
-    public void setLabels(Set<String> labels) {
-        this.labels = labels;
-    }
-    
-    public String getExpression() {
-        return expression;
-    }
-    
-    public void setExpression(String expression) {
-        this.expression = expression;
-    }
-    
-    @Override
-    protected List<T> doSelect(CmdbContext<T> context) {
-        if (CollectionUtils.isEmpty(labels)) {
-            return context.getProviders()
-                    .stream()
-                    .map(CmdbContext.CmdbInstance::getInstance)
-                    .collect(Collectors.toList());
-        }
-        CmdbContext.CmdbInstance<T> consumer = context.getConsumer();
-        Map<String, String> consumerLabels = Optional.ofNullable(consumer.getEntity())
-                .map(Entity::getLabels)
-                .orElse(Collections.emptyMap());
-        
-        // filter the instance if consumer and providers' label values equals.
-        List<T> result = context.getProviders()
-                .stream()
-                .filter(ci -> {
-                    Entity providerEntity = ci.getEntity();
-                    if (Objects.isNull(providerEntity)) {
-                        return false;
-                    }
-                    Map<String, String> providerLabels = Optional.ofNullable(ci.getEntity().getLabels())
-                            .orElse(Collections.emptyMap());
-                    return labels.stream()
-                            .allMatch(label -> {
-                                String consumerLabelValue = consumerLabels.get(label);
-                                if (StringUtils.isBlank(consumerLabelValue)) {
-                                    return false;
-                                }
-                                return Objects.equals(consumerLabelValue, providerLabels.get(label));
-                            });
-                })
-                .map(CmdbContext.CmdbInstance::getInstance)
-                .collect(Collectors.toList());
-        
-        // if none match, then return all providers.
-        if (CollectionUtils.isEmpty(result)) {
-            return context.getProviders()
-                    .stream()
-                    .map(CmdbContext.CmdbInstance::getInstance)
-                    .collect(Collectors.toList());
-        }
-        return result;
-    }
-    
-    @Override
-    public Selector<List<T>, CmdbContext<T>, String> parse(String condition) throws NacosException {
-        this.labels = LabelSelector.ExpressionInterpreter.parseExpression(condition);
-        this.expression = condition;
-        return this;
-    }
-    
-    @Override
-    public String getType() {
-        return TYPE;
-    }
-}
