+++ naming/src/main/java/com/alibaba/nacos/naming/selector/context/CmdbSelectorContextBuilder.java
@@ -20,8 +20,8 @@ package com.alibaba.nacos.naming.selector.context;
 import com.alibaba.nacos.api.cmdb.pojo.Entity;
 import com.alibaba.nacos.api.cmdb.pojo.PreservedEntityTypes;
 import com.alibaba.nacos.api.naming.pojo.Instance;
-import com.alibaba.nacos.api.naming.selector.context.CmdbContext;
-import com.alibaba.nacos.api.naming.selector.context.SelectorContextBuilder;
+import com.alibaba.nacos.api.selector.context.CmdbContext;
+import com.alibaba.nacos.api.selector.context.SelectorContextBuilder;
 import com.alibaba.nacos.cmdb.service.CmdbReader;
 import com.alibaba.nacos.sys.utils.ApplicationUtils;
