+++ naming/src/main/resources/META-INF/services/com.alibaba.nacos.api.selector.Selector
@@ -15,5 +15,5 @@
 #
 #
 
-com.alibaba.nacos.naming.selector.CmdbLabelSelector
-com.alibaba.nacos.naming.selector.NoneRemoveSelector
+com.alibaba.nacos.naming.selector.LabelSelector
+com.alibaba.nacos.naming.selector.NoneSelector
